Simplify

Documentation

simplify :: Condition -> Condition

Simplifies a condition for legibility and stack pointer matching correctness.

simplifyBoolean :: Condition -> Condition

Rewrite conditions when we can know their results, increasing readability.

simplifyArithmetic :: Expr -> Expr

Simplify additions and subtractions to their canonical form. This is nice to create readable output, but essential for matching the stack pointer!