Test

Documentation

max4Example :: (Condition, Condition, [Instr])

See max4.lang0.

sqrtExample :: (Condition, Condition, [Instr])

The floor of the square root function: return(floor(sqrt(arg0))). See sqrt.lang0.

example3 :: (Condition, Condition, [Instr])

A rather complicated way of saying return(a0), by incrementing l0 to the proper value step-by-step and then returning it. This is useful to trick the simplifier, it cannot oversimplify this kind of programs.

isEvenExample :: (Condition, Condition, [Instr])

Tests whether a natural number is even. See isEven.lang0.

test :: (t, Condition, [Instr]) -> IO ()

Debugging tool for programs. Prints the weakest precondition of each instruction in the program.