ToSBV

Description

Contains functionality for converting our immediate representation of predicates (the Condition datatype) into an SBV representation.

Documentation

data IllegalVariableException

Thrown whenever a variable occurs within a condition that is not supposed to be present in a pre- or post-condition.

preConditionToPredicate :: Quantifier -> Condition -> Predicate

Transforms the intermediate representation of a (weakest) precondition into a Predicate as understood by SBV, quantifying over stale variable with the given quantifier.