WP

Documentation

wp :: Sequence -> Condition -> Condition

Gives the weakest precondition of the given program, given the postcondition.

wps :: Sequence -> Condition -> [Condition]

Gives the sequence of weakest preconditions for each instruction of the given program, given the postcondition.